+++ confidant/routes/certificates.py
@@ -4,6 +4,7 @@ from flask import blueprints, jsonify, request
 
 from confidant import authnz, settings
 from confidant.services.certificates import certificatemanager
+from confidant.services.certificates import certificate_authority
 from confidant.schema.certificates import (
     certificate_authority_response_schema,
     certificate_authorities_response_schema,
@@ -16,15 +17,15 @@ from confidant.schema.certificates import (
 from confidant.utils import misc
 
 logger = logging.getLogger(__name__)
-blueprint = blueprints.Blueprint('certificates', __name__)
+blueprint = blueprints.Blueprint("certificates", __name__)
 
 acl_module_check = misc.load_module(settings.ACL_MODULE)
 
 
-@blueprint.route('/v1/certificates/<ca>/<cn>', methods=['GET'])
+@blueprint.route("/v1/certificates/<ca>/<cn>", methods=["GET"])
 @authnz.require_auth
 def get_certificate(ca, cn):
-    '''
+    """
     Get a certificate from the provided CA, for the provided CN.
 
     .. :quickref: Certificate; Get certificate from the provided CA, for the
@@ -66,34 +67,35 @@ def get_certificate(ca, cn):
     :statuscode 200: success
     :statuscode 403: client does not have access to generate the requested
                      certificate.
-    '''
+    """
     try:
         ca_object = certificatemanager.get_ca(ca)
     except certificatemanager.CertificateAuthorityNotFoundError:
-        return jsonify({'error': 'Provided CA not found.'}), 404
-    san = request.args.getlist('san')
+        return jsonify({"error": "Provided CA not found."}), 404
+    san = request.args.getlist("san")
 
     logged_in_user = authnz.get_logged_in_user()
     if not acl_module_check(
-        resource_type='certificate',
-        action='get',
+        resource_type="certificate",
+        action="get",
         resource_id=cn,
         kwargs={
-            'ca': ca,
-            'san': san,
+            "ca": ca,
+            "san": san,
         },
     ):
-        msg = ('{} does not have access to get certificate cn {} against'
-               ' ca {}').format(
+        msg = (
+            "{} does not have access to get certificate cn {} against" " ca {}"
+        ).format(
             authnz.get_logged_in_user(),
             cn,
             ca,
         )
-        error_msg = {'error': msg, 'reference': cn}
+        error_msg = {"error": msg, "reference": cn}
         return jsonify(error_msg), 403
 
     logger.info(
-        'get_certificate called on id={} for ca={} by user={}'.format(
+        "get_certificate called on id={} for ca={} by user={}".format(
             cn,
             ca,
             logged_in_user,
@@ -101,8 +103,8 @@ def get_certificate(ca, cn):
     )
 
     validity = request.args.get(
-        'validity',
-        default=ca_object.settings['max_validity_days'],
+        "validity",
+        default=ca_object.settings["max_validity_days"],
         type=int,
     )
     try:
@@ -111,25 +113,25 @@ def get_certificate(ca, cn):
             validity,
             san,
         )
-    except certificatemanager.CertificateNotReadyError:
+    except certificate_authority.CertificateNotReadyError:
         # Ratelimit response for a locked certificate in the cache
-        error_msg = 'Certificate being requested, please wait and try again.'
+        error_msg = "Certificate being requested, please wait and try again."
         response = jsonify(error_msg)
         response.retry_after = 2
         return response, 429
     certificate_response = CertificateResponse(
-        certificate=certificate['certificate'],
-        certificate_chain=certificate['certificate_chain'],
-        key=certificate['key'],
+        certificate=certificate["certificate"],
+        certificate_chain=certificate["certificate_chain"],
+        key=certificate["key"],
     )
     return certificate_expanded_response_schema.dumps(certificate_response)
 
 
-@blueprint.route('/v1/certificates/<ca>', methods=['POST'])
+@blueprint.route("/v1/certificates/<ca>", methods=["POST"])
 @authnz.require_auth
 @authnz.require_csrf_token
 def get_certificate_from_csr(ca):
-    '''
+    """
     Get a certificate from the ca provided in the url, using the CSR, validity
     and san provided in the POST body.
 
@@ -170,27 +172,33 @@ def get_certificate_from_csr(ca):
                      or was missing from the request.
     :statuscode 403: Client does not have access to generate the requested
                      certificate.
-    '''
+    """
     try:
         ca_object = certificatemanager.get_ca(ca)
-    except certificatemanager.CertificateAuthorityNotFoundError:
-        return jsonify({'error': 'Provided CA not found.'}), 404
+    except certificate_authority.CertificateAuthorityNotFoundError:
+        return jsonify({"error": "Provided CA not found."}), 404
     data = request.get_json()
-    if not data or not data.get('csr'):
-        return jsonify(
-            {'error': 'csr must be provided in the POST body.'},
-        ), 400
+    if not data or not data.get("csr"):
+        return (
+            jsonify(
+                {"error": "csr must be provided in the POST body."},
+            ),
+            400,
+        )
     validity = data.get(
-        'validity',
-        ca_object.settings['max_validity_days'],
+        "validity",
+        ca_object.settings["max_validity_days"],
     )
     try:
-        csr = ca_object.decode_csr(data['csr'])
+        csr = ca_object.decode_csr(data["csr"])
     except Exception:
-        logger.exception('Failed to decode PEM csr')
-        return jsonify(
-            {'error': 'csr could not be decoded'},
-        ), 400
+        logger.exception("Failed to decode PEM csr")
+        return (
+            jsonify(
+                {"error": "csr could not be decoded"},
+            ),
+            400,
+        )
     # Get the cn and san values from the csr object, so that we can use them
     # for the ACL check.
     cn = ca_object.get_csr_common_name(csr)
@@ -198,43 +206,44 @@ def get_certificate_from_csr(ca):
 
     logged_in_user = authnz.get_logged_in_user()
     if not acl_module_check(
-        resource_type='certificate',
-        action='get',
+        resource_type="certificate",
+        action="get",
         resource_id=cn,
         kwargs={
-            'ca': ca,
-            'san': san,
+            "ca": ca,
+            "san": san,
         },
     ):
-        msg = ('{} does not have access to get certificate cn {} against'
-               ' ca {}').format(
+        msg = (
+            "{} does not have access to get certificate cn {} against" " ca {}"
+        ).format(
             authnz.get_logged_in_user(),
             cn,
             ca,
         )
-        error_msg = {'error': msg, 'reference': cn}
+        error_msg = {"error": msg, "reference": cn}
         return jsonify(error_msg), 403
 
     logger.info(
-        'get_certificate called on id={} for ca={} by user={}'.format(
+        "get_certificate called on id={} for ca={} by user={}".format(
             cn,
             ca,
             logged_in_user,
         )
     )
 
-    certificate = ca_object.issue_certificate(data['csr'], validity)
+    certificate = ca_object.issue_certificate(data["csr"], validity)
     certificate_response = CertificateResponse(
-        certificate=certificate['certificate'],
-        certificate_chain=certificate['certificate_chain'],
+        certificate=certificate["certificate"],
+        certificate_chain=certificate["certificate_chain"],
     )
     return certificate_response_schema.dumps(certificate_response)
 
 
-@blueprint.route('/v1/cas', methods=['GET'])
+@blueprint.route("/v1/cas", methods=["GET"])
 @authnz.require_auth
 def list_cas():
-    '''
+    """
     List the configured CAs.
 
     .. :quickref: Certificate Authorities; Get a list of the detailed
@@ -268,31 +277,31 @@ def list_cas():
     :resheader Content-Type: application/json
     :statuscode 200: success
     :statuscode 403: client does not have access to list CAs
-    '''
+    """
 
     logged_in_user = authnz.get_logged_in_user()
     if not acl_module_check(
-        resource_type='ca',
-        action='list',
+        resource_type="ca",
+        action="list",
     ):
-        msg = '{} does not have access to list cas'.format(
+        msg = "{} does not have access to list cas".format(
             authnz.get_logged_in_user(),
         )
-        error_msg = {'error': msg}
+        error_msg = {"error": msg}
         return jsonify(error_msg), 403
 
     cas = certificatemanager.list_cas()
 
-    logger.info('list_cas called by user={}'.format(logged_in_user))
+    logger.info("list_cas called by user={}".format(logged_in_user))
 
     cas_response = CertificateAuthoritiesResponse.from_cas(cas)
     return certificate_authorities_response_schema.dumps(cas_response)
 
 
-@blueprint.route('/v1/cas/<ca>', methods=['GET'])
+@blueprint.route("/v1/cas/<ca>", methods=["GET"])
 @authnz.require_auth
 def get_ca(ca):
-    '''
+    """
     Get the CA information for the provided ca.
 
     .. :quickref: Certificate Authorities; Get the detailed certificate
@@ -327,27 +336,27 @@ def get_ca(ca):
     :resheader Content-Type: application/json
     :statuscode 200: Success
     :statuscode 403: Client does not have access to get the requested CA.
-    '''
+    """
     try:
         ca_object = certificatemanager.get_ca(ca)
-    except certificatemanager.CertificateAuthorityNotFoundError:
-        return jsonify({'error': 'Provided CA not found.'}), 404
+    except certificate_authority.CertificateAuthorityNotFoundError:
+        return jsonify({"error": "Provided CA not found."}), 404
 
     logged_in_user = authnz.get_logged_in_user()
     if not acl_module_check(
-        resource_type='ca',
-        action='get',
+        resource_type="ca",
+        action="get",
         resource_id=ca,
     ):
-        msg = '{} does not have access to get ca {}'.format(
+        msg = "{} does not have access to get ca {}".format(
             authnz.get_logged_in_user(),
             ca,
         )
-        error_msg = {'error': msg, 'reference': ca}
+        error_msg = {"error": msg, "reference": ca}
         return jsonify(error_msg), 403
 
     logger.info(
-        'get_ca called on id={} by user={}'.format(
+        "get_ca called on id={} by user={}".format(
             ca,
             logged_in_user,
         )
@@ -355,9 +364,9 @@ def get_ca(ca):
 
     _ca = ca_object.get_certificate_authority_certificate()
     ca_response = CertificateAuthorityResponse(
-        ca=_ca['ca'],
-        certificate=_ca['certificate'],
-        certificate_chain=_ca['certificate_chain'],
-        tags=_ca['tags'],
+        ca=_ca["ca"],
+        certificate=_ca["certificate"],
+        certificate_chain=_ca["certificate_chain"],
+        tags=_ca["tags"],
     )
     return certificate_authority_response_schema.dumps(ca_response)
